requirements.txt
python-telegram-bot
langchain-groq
langchain-community
langchain-core
langchain
sentence-transformers
faiss-cpu
groq
python-dotenv
nest_asyncio